Digital Trigger Style
Added a new setting to change the behavior of the trigger buttons from
analog (depending on the amount of pressure send corresponding value
from 0.0 to 1.0) to digital (send 1.0 as soon as the trigger is pressed
more than 0%).

Many games I tried needed the triggers to be pressed for 100% which was
a pain.