- WIP iOS hardware keyboard responsiveness updates for games where long down states are used for interactions   with proposed api for keyboard profiles* - adjust left click drag time - added support for right click drag behavior
*
The problem this solves is responsiveness from held down keypresses.
We don't get down states from hardware keyboards (now atleast).
This attempts to circumvent the issue by making those sensitive keys toggable in the interim.
This issue will go away if apple exposes that in a future api update, at which point we can decide to either
keep these profiles, or deprecate them. This was designed to be more customizable than just toggable, but for now
thats all I've chosen to implement.