Add support for changing the sensitivity of relative mouse mode.
This commit introduces support on the iOS and tvOS apps for
changing the mouse sensitivity within relative mouse ("Touchpad")
mode. It does this by multiplying the delta movement sent to the
host by a fixed factor which can be configured by the user in Settings.

On iOS, the setting provided is a slider with non-lineal behavior.
While it has a multiplier range of 0.0 to 6.0, the default value
of 1.0 stays in the middle; thus, the left side comprehends a
range of [0.0, 1.0) and the right side uses a different range of
[1.0, 6.0]. The option of having 0x sensibility is explicitely
provided in order to unlock usecases in which locking the mouse
cursor in a fixed place is desired.

In addition to that, the slider has a soft detent on the
default 1.0x setting, which allows to quickly reset the setting
to a default state. This is accompanied by haptic feedback for
better UX.

On tvOS, due to the lack of a slider control support in Settings.bundle,
a multi-input is provided instead that roughly covers the iOS's
slider range.

A Core Data migration and accompanying glue code is provided for
persisting the setting change.